Fix Ambiguous Audio Track Selection
This addresses #38 where there are multiple audio tracks with the title
"English”, and when trying to select different ones, there is no change in the
audio track that is selected.

### Cause

The issue is a byproduct of how F1TV declares the various audio tracks in the
`m3u8`, the available metadata from AVFoundation, and the way in which the app
currently selects which media track.

“Good” `m3u8` example (user info redacted):

```other
#EXT-X-MEDIA:TYPE=AUDIO,GROUP-ID="Encoding_0",NAME="English",DEFAULT=YES,AUTOSELECT=YES,LANGUAGE="eng",URI=".../OTT_F1_WIF_AUDIO_1_64000_eng_vod.m3u8"
#EXT-X-MEDIA:TYPE=AUDIO,GROUP-ID="Encoding_0",NAME="Deutsch",AUTOSELECT=YES,LANGUAGE="deu",URI=".../OTT_F1_WIF_AUDIO_2_64000_ger_vod.m3u8"
#EXT-X-MEDIA:TYPE=AUDIO,GROUP-ID="Encoding_0",NAME="Francais",AUTOSELECT=YES,LANGUAGE="fra",URI=".../OTT_F1_WIF_AUDIO_3_64000_fre_vod.m3u8"
#EXT-X-MEDIA:TYPE=AUDIO,GROUP-ID="Encoding_0",NAME="Espanol",AUTOSELECT=YES,LANGUAGE="spa",URI=".../OTT_F1_WIF_AUDIO_4_64000_spa_vod.m3u8"
#EXT-X-MEDIA:TYPE=AUDIO,GROUP-ID="Encoding_0",NAME="FX",AUTOSELECT=YES,LANGUAGE="cfx",URI=".../OTT_F1_WIF_AUDIO_0_64000_eng_vod.m3u8"
```

“Bad” m3u8 example (user info redacted):

```other
#EXT-X-MEDIA:TYPE=AUDIO,GROUP-ID="Encoding_0",NAME="FX",DEFAULT=YES,AUTOSELECT=YES,LANGUAGE="eng",URI=".../OTT_F1_WIF_AUDIO_0_64000_eng_vod.m3u8"
#EXT-X-MEDIA:TYPE=AUDIO,GROUP-ID="Encoding_0",NAME="English",AUTOSELECT=YES,LANGUAGE="eng",URI=".../OTT_F1_WIF_AUDIO_1_64000_eng_vod.m3u8"
#EXT-X-MEDIA:TYPE=AUDIO,GROUP-ID="Encoding_0",NAME="German",AUTOSELECT=YES,LANGUAGE="ger",URI=".../OTT_F1_WIF_AUDIO_2_64000_ger_vod.m3u8"
#EXT-X-MEDIA:TYPE=AUDIO,GROUP-ID="Encoding_0",NAME="French",AUTOSELECT=YES,LANGUAGE="fre",URI=".../OTT_F1_WIF_AUDIO_3_64000_fre_vod.m3u8"
#EXT-X-MEDIA:TYPE=AUDIO,GROUP-ID="Encoding_0",NAME="Spanish",AUTOSELECT=YES,LANGUAGE="spa",URI=".../OTT_F1_WIF_AUDIO_4_64000_spa_vod.m3u8"
```

If you look closely at the “FX” track, in the bad example it has a `LANGUAGE` of
`eng`. It has a unique `Name`, so it should possible to disambiguate the audio
streams. However, the properties surfaced by AVFoundation – specifically
`displayName` – don't directly match to these. `displayName` is actually a
localised (by the device) string derived from the `LANGUAGE` attribute. This
means in the ‘bad' scenario, there are two items in the audio tracks with a
`displayName` of "English”.

The logic used by the app to select the specific track is driven by searching
for a matching `displayName` in the from the assets auto tracks (sourced via
`availableMediaCharacteristicsWithMediaSelectionOptions`,`AVMediaSelectionGroup`,
and `AVMediaSelectionOption`). But since there are multiple items with the
*same* `displayName`, the search matches the first time, every time.

> For more details on the display name behaviours, see [this
post](https://developer.apple.com/forums/thread/650968) from Apple.

Of interest, if you spelunk through the objects returns by the APIs, you can
actually see the `Name` attributes value in a private dictionary, who's key
appears only in [WebKit source
code](https://github.com/WebKit/WebKit/blob/main/Tools/TestWebKitAPI/Tests/WebCore/cocoa/AVFoundationSoftLinkTest.mm#L123)
for managing & implementing MSE.

### Fix

`AVMediaSelectionOption` has a
[method](https://developer.apple.com/documentation/avfoundation/avmediaselectionoption/1386310-propertylist)
that returns a property list that, per apple:

> … that’s sufficient to identify the option within its group.

When used with `AVMediaSelectionGroup.mediaSelectionGroup(withPropertyList:)`,
and `AVPlayerItem.select(AVMediaSelectionOption?, in:)`, you can uniquely
identify the items – even with the duplicate `displayName`.

With this in mind, the extensions in `AVPlayerItemExtension` are updated to
return & consume a new type (`MediaTrackDto`) that holds a display name
(displayed to the user), and the result of calling
`AVMediaSelectionOption.propertyList`. Maintaining the same calling patterns,
with the additional data, everything functions great 🎉.

Additionally `displayName` was used in persisting the preferred language setting
when the audio / caption language was changed. Since I don't believe that the
property list is reusable across different playlists, the existing mechanism
using `displayName` was left in place.