"Add channel" button in control strip
Adds a button next to the "remove channel" button in the player control
strip that hides the control strip, then opens the channel listing.
This allows adding new channels to the player grid without needing to
use a left swipe gesture on a remote or phone. This is useful for the
new Siri Remote, where users can disable the swipe gestures, or for
alternative inputs like game controllers or universal remotes that do
not support swipe gestures.

Tested with ATV4K simulator build, and was able to add three other
channels just using arrow keys, enter, and escape (dpad, select, and
back respectively).

Fixes #33